feat(rendering): add shadow mapping and new shaders for improved lighting
feat(audio): implement audio type system and add placeholder assets

feat(networking): add network message system for multiplayer support

feat(world): enhance chunk lighting with propagation algorithm

feat(entities): create base entity system for game objects

refactor(physics): clean up ocean physics implementation

refactor(terrain): improve terrain generator logging

fix(world): fix trade route waypoint generation randomness

fix(noise): correct floating point precision in noise generator

fix(camera): remove unused quaternion import

style(rendering): reformat shader code for better readability

docs(audio): add documentation for audio resource structure

chore(build): update compiler settings and maven configuration

chore(config): add controller settings to default properties

chore(ide): update vscode settings for maven hierarchical view

test(rendering): add shadow map rendering tests

ci: update compiler plugin configuration for Java 25

perf(rendering): optimize shadow map rendering pipeline

perf(world): optimize block light propagation in chunks

+++ src/main/java/com/odyssey/world/TradeRoute.java
@@ -108,8 +108,8 @@ public class TradeRoute {
                 );
                 
                 // Add some variation to avoid straight lines
-                waypoint.x += (Math.random() - 0.5) * 100.0f;
-                waypoint.y += (Math.random() - 0.5) * 100.0f;
+                waypoint.x += (float)((Math.random() - 0.5) * 100.0);
+                waypoint.y += (float)((Math.random() - 0.5) * 100.0);
                 
                 waypoints.add(waypoint);
             }
