+++ src/main/java/com/odyssey/world/noise/NoiseGenerator.java
@@ -119,8 +119,8 @@ public class NoiseGenerator {
         int X = (int) Math.floor(x) & 255;
         int Y = (int) Math.floor(y) & 255;
         
-        x -= Math.floor(x);
-        y -= Math.floor(y);
+        x -= (float)Math.floor(x);
+        y -= (float)Math.floor(y);
         
         float u = fade(x);
         float v = fade(y);
@@ -142,9 +142,9 @@ public class NoiseGenerator {
         int Y = (int) Math.floor(y) & 255;
         int Z = (int) Math.floor(z) & 255;
         
-        x -= Math.floor(x);
-        y -= Math.floor(y);
-        z -= Math.floor(z);
+        x -= (float)Math.floor(x);
+        y -= (float)Math.floor(y);
+        z -= (float)Math.floor(z);
         
         float u = fade(x);
         float v = fade(y);
